+++ src/main/java/eu/xenit/gradle/docker/alfresco/tasks/LabelConsumerTask.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import java.util.function.Supplier;
 import org.gradle.api.Task;
 import org.gradle.api.provider.Provider;
+import org.gradle.api.tasks.TaskProvider;
 
 public interface LabelConsumerTask extends Task, LabelConsumerExtension {
 
@@ -21,5 +22,11 @@ public interface LabelConsumerTask extends Task, LabelConsumerExtension {
     default void withLabels(LabelSupplierTask task) {
         dependsOn(task);
         withLabels(task.getLabels());
+    }
+
+    @Override
+    default void withLabels(TaskProvider<? extends LabelSupplierTask> taskProvider) {
+        dependsOn(taskProvider);
+        withLabels(taskProvider.flatMap(LabelSupplierTask::getLabels));
     }
 }