+++ src/main/java/eu/xenit/gradle/docker/alfresco/tasks/extension/DockerfileWithWarsExtension.java
@@ -2,6 +2,7 @@ package eu.xenit.gradle.docker.alfresco.tasks.extension;
 
 import com.bmuschko.gradle.docker.tasks.image.Dockerfile;
 import eu.xenit.gradle.docker.alfresco.tasks.WarLabelOutputTask;
+import eu.xenit.gradle.docker.alfresco.tasks.WarOutputTask;
 import java.io.File;
 import java.util.function.Supplier;
 import org.gradle.api.artifacts.Configuration;
@@ -9,6 +10,7 @@ import org.gradle.api.file.RegularFile;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.tasks.Input;
+import org.gradle.api.tasks.TaskProvider;
 
 public interface DockerfileWithWarsExtension {
 
@@ -37,6 +39,10 @@ public interface DockerfileWithWarsExtension {
     void addWar(String name, Supplier<File> file);
 
     void addWar(String name, Configuration configuration);
+
+    default void addWar(String name, TaskProvider<? extends WarLabelOutputTask> taskProvider) {
+        addWar(name, taskProvider.flatMap(WarOutputTask::getOutputWar));
+    }
 
     static DockerfileWithWarsExtension get(Dockerfile task) {
         return task.getConvention().getPlugin(DockerfileWithWarsExtension.class);